+++ BearishEngine/src/Graphics/BETFile.h
@@ -78,7 +78,7 @@ namespace Bearish { namespace Graphics {
 
 			free(data);
 			data = (u8*)malloc(colors.size() * 5);
-			dataSize = colors.size() * 5;
+			dataSize = (u32)colors.size() * 5;
 
 			pos = 0;
 			for (auto& pair : colors) {
